LayoutTests/http/tests/security/shape-image-cors-redirect-error-message-logging-1-expected.txt

CONSOLE MESSAGE: Unsafe attempt to load URL http://127.0.0.1:8000/resources/redirect.php?url=http%3A%2F%2Flocalhost%3A8080%2Fsecurity%2Fresources%2Fimage-access-control.php%3Ffile%3D..%2F..%2Fresources%2Fsquare100.png%26allow%3Dfalse.
Verify the error message in console in case of CORS failing checks.
